requirements.txt
python-telegram-bot
google-generativeai
yfinance
python-dotenv
schedule
nest_asyncio
